Add very basic Targa reading to plClientResMgr.
The Targa reader only currently supports the image type used by
Plasma 2.1+ games, so we need to make sure we can skip to the end
of the resource (or continue on if the resource isn't a valid Targa
at all).

All Plasma 2.1+ games using the Cursors.dat file appear to have their
resources named with the .tga extension, so I've moved the Targa
reader and writer into their own block, leaving unsupported types
separate.  We don't want to try reading nonsense as a Targa image.