TEMP: This commit is a bullet WIP.  DO NOT USE.
This was an attempt to reincorporate the changes made since the original work
on this branch before it was rebased, over-writing optimizations and cleanup
made by branan, skoader, and others.  It is incomplete and WILL COMPILE BUT
CRASHES WHEN RUN, and is included for reference.

DO NOT USE THIS COMMIT AS-IS for future work.